fix: add error message
