+++ app/error.styled.tsx
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+	height: 100%;
+	display: flex;
+	align-items: center;
+	flex-direction: column;
+	padding: 32px;
+`;
+
+export const Content = styled.div`
+	margin-top: 24px;
+	text-align: center;
+`;
