+++ app/globals.css
@@ -4,6 +4,7 @@ body {
 	margin: 0;
 	font-family: "Roboto", -apple-system, BlinkMacSystemFont, Segoe UI, Roboto,
 		Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+	height: 100vh;
 }
 
 a {
