+++ app/page.tsx
@@ -1,27 +1,5 @@
-import { cookies } from "next/headers";
-
-import YAML from "yaml";
 import { OpenapiDoc } from "@components/OpenapiDoc";
-
-async function getData({
-	name,
-	namespace,
-}: {
-	name: string;
-	namespace: string;
-}) {
-	const nextCookies = cookies();
-
-	if (!name || !namespace) {
-		throw Error("Cannot fetch data without name and namespace");
-	}
-	const res = await fetch(
-		`${process.env.kuskAPI}/apis/${namespace}/${name}/crd`
-	);
-
-	const data = await res.json();
-	return YAML.parse(data?.spec?.spec);
-}
+import { getData } from "./data";
 
 export default async function Home({
 	searchParams,
