+++ next.config.js
@@ -5,9 +5,6 @@ const nextConfig = {
 		fontLoaders: [
 			{
 				loader: "@next/font/google",
-				weight: ["400", "500", "700", "900"],
-				subsets: ["latin"],
-				preload: true,
 			},
 		],
 	},
