+++ package.json
@@ -24,6 +24,7 @@
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
 		"react-redux": "^8.0.4",
+		"server-only": "^0.0.1",
 		"sort-by": "^0.0.2",
 		"styled-components": "^5.3.6",
 		"typescript": "4.9.3",
